+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Changelog:
 	- Fixed cpdos_main: Reworking the source code to avoid FP and improve detection, as well as being able to send headers not authorized by the basic requests library & recreating a “fresh” session before launching the cpdos modules
 	- New payloads
 	- Fixed logic and style bug
+	News:
+	- CVE-2025-57822 module check
 ---------
 
 - 2.0